fix(focus): app doesn't work on flutter apps
