+++ app/src/main/java/com/neo/speaktouch/controller/FocusController.kt
@@ -22,8 +22,8 @@ import android.view.accessibility.AccessibilityNodeInfo
 import com.neo.speaktouch.intercepter.event.CallbackInterceptor
 import com.neo.speaktouch.model.NodeFilter
 import com.neo.speaktouch.utils.extension.Direction
-import com.neo.speaktouch.utils.extension.performFocus
 import com.neo.speaktouch.utils.extension.nodeScan
+import com.neo.speaktouch.utils.extension.performFocus
 import dagger.hilt.android.scopes.ServiceScoped
 import javax.inject.Inject
