+++ app/src/main/java/com/neo/speaktouch/model/NodeFilter.kt
@@ -30,7 +30,7 @@ sealed interface NodeFilter {
         override fun filter(node: AccessibilityNodeInfo): Boolean {
             val compat = AccessibilityNodeInfoCompat.wrap(node)
 
-            if (!NodeValidator.isValidForAccessibility(compat)) return false
+            if (!compat.isVisibleToUser) return false
 
             if (NodeValidator.mustFocus(compat)) return true
