+++ app/src/main/java/com/neo/speaktouch/utils/Node.kt
@@ -20,11 +20,11 @@ package com.neo.speaktouch.utils
 
 import androidx.core.view.accessibility.AccessibilityNodeInfoCompat
 import com.neo.speaktouch.utils.extension.getContent
+import com.neo.speaktouch.utils.extension.getName
 import com.neo.speaktouch.utils.extension.iterator
 import com.neo.speaktouch.utils.extension.map
 import org.json.JSONArray
 import org.json.JSONObject
-import java.lang.reflect.Method
 
 data class Node(
     val nodeInfo: AccessibilityNodeInfoCompat,
@@ -37,6 +37,7 @@ data class Node(
 
         json.put("content", content)
         json.put("className", nodeInfo.className)
+        json.put("isVisibleToUser", nodeInfo.isVisibleToUser)
         json.put("isImportantForAccessibility", nodeInfo.isImportantForAccessibility)
         json.put("actions", nodeInfo.actionList.map { it.getName() })
 
@@ -50,17 +51,4 @@ data class Node(
 
         return json
     }
-}
-
-fun AccessibilityNodeInfoCompat.AccessibilityActionCompat.getName(): String {
-    val clazz = AccessibilityNodeInfoCompat::class.java
-
-    val method: Method = clazz.getDeclaredMethod(
-        "getActionSymbolicName",
-        Int::class.javaPrimitiveType
-    )
-
-    method.isAccessible = true
-
-    return method.invoke(null, id) as String
 }
